+++ .eslintrc.cjs
@@ -25,7 +25,8 @@ module.exports = {
     '@typescript-eslint/no-non-null-assertion': 'off',
     '@typescript-eslint/no-var-requires': 'off',
     'vue/require-default-prop': 'off',
-    'vue/multi-word-component-names': 'off'
+    'vue/multi-word-component-names': 'off',
+    "@typescript-eslint/explicit-function-return-type": "off"
   },
   overrides: [
     {
