+++ electron.vite.config.ts
@@ -1,6 +1,7 @@
 import { resolve } from 'path'
 import { defineConfig, externalizeDepsPlugin } from 'electron-vite'
 import vue from '@vitejs/plugin-vue'
+import VueRouter from "unplugin-vue-router/vite";
 
 export default defineConfig({
   main: {
@@ -31,11 +32,22 @@ export default defineConfig({
     }
   },
   renderer: {
+    build: {
+      watch: {},
+      rollupOptions: {
+        input: {
+          index: resolve(__dirname, 'src/renderer/renderer-main.ts')
+        }
+      }
+    },
     resolve: {
       alias: {
         '@renderer': resolve('src/renderer/src')
       }
     },
-    plugins: [vue()]
+    plugins: [VueRouter({
+      routesFolder: 'src/renderer/src/pages',
+      dts: 'src/renderer/typed-router.d.ts',
+    }),vue(),]
   }
 })
