+++ package.json
@@ -28,7 +28,8 @@
   "dependencies": {
     "@electron-toolkit/preload": "^1.0.3",
     "@electron-toolkit/utils": "^1.0.2",
-    "electron-updater": "^5.3.0"
+    "electron-updater": "^5.3.0",
+    "vue-router": "^4.2.4"
   },
   "devDependencies": {
     "@electron-toolkit/tsconfig": "^1.0.1",
@@ -52,6 +53,7 @@
     "prettier": "^2.8.7",
     "sass": "^1.62.1",
     "typescript": "^4.9.5",
+    "unplugin-vue-router": "^0.6.4",
     "vite": "^4.2.1",
     "vue": "^3.2.47",
     "vue-tsc": "^1.2.0"
