+++ src/renderer/index.html
@@ -27,6 +27,6 @@
 
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="/renderer-main.ts"></script>
   </body>
 </html>
