--- src/renderer/src/main.ts
@@ -1,3 +1,0 @@
-import { createApp } from 'vue'
-import App from './App.vue'
-createApp(App).mount('#app')
