+++ src/renderer/src/pages/option.vue
@@ -1,0 +1,9 @@
+<template>
+  option
+</template>
+
+<script lang="ts">
+export default {
+
+}
+</script>
