+++ src/renderer/tsconfig.web.json
@@ -4,8 +4,11 @@
     "src/renderer/src/env.d.ts",
     "src/renderer/src/**/*",
     "src/renderer/src/**/*.vue",
-    "src/preload/*.d.ts"
+    "src/preload/*.d.ts",
+    "src/renderer/typed-router.d.ts"
+
   ],
+  "exclude": ["vue-router"],
   "compilerOptions": {
     "composite": true,
     "esModuleInterop": true,
