+++ tsconfig.json
@@ -2,6 +2,5 @@
   "files": [],
   "references": [{ "path": "src/main/tsconfig.node.json" }, { "path": "src/renderer/tsconfig.web.json" }],
   "rules": {
-    "@typescript-eslint/explicit-function-return-type": 0
   }
 }
